fix(porting_layer): move asprintf to standalone file
