+++ src/porting_layer/src/asprintf.h
@@ -1,0 +1,4 @@
+#ifndef HAVE_ASPRINTF
+#define HAVE_ASPRINTF
+int asprintf(char **strp, const char *fmt, ...);
+#endif
